Removes test that doesn't replicate behavior from SW.
This test does a CJK query and looks for vernacular fields. The query
at one point returned only multi-script records but that is no longer
the case, thus the test fails. The search terms are in Latin script and
would not be sent as a CJK query from SearchWorks anyways.