Modifies our "OR" boolean tests to match SearchWorks queries
Our previous practice in sw_index_tests relied on Solr 4.x boolean logic
but was not actually testing the queries that SearchWorks executes when
it comes to "OR" boolean logic. This commit fixes that by using the
correct query parser and providing the mm values that are modified
by the Blacklight Advanced Search plugin.

Fixes #193, Fixes #191